refactor(prd-admin): 前端搭档视觉对齐视觉创作智能体靛紫全息风格
