+++ changelogs/2026-06-10_front-end-agent-layout-refresh.md
@@ -2,6 +2,8 @@
 | feat | prd-admin | 截图视觉诊断支持 Ctrl+V 粘贴截图、预览与上传，随请求发送 Vision 识图 |
 | feat | prd-admin | 粒子宇宙凄凉背景（星尘/流星/冷色星云）+ 按钮手型悬停抬起与按下反馈 |
 | feat | prd-admin | 怀旧星夜视觉：暖色标题/胶片颗粒/扫描线、密集流星与暖色玻璃面板 |
+| refactor | prd-admin | 前端搭档视觉对齐视觉创作智能体：夜景+粒子漩涡、全息标题、靛紫玻璃面板 |
+| refactor | prd-admin | 抽取共享 NightSkyBackground 组件供视觉/前端搭档复用 |
 | fix | prd-admin | 目标技术栈默认改为 Vue3 + UniApp 小程序 |
 | feat | prd-api | 前端搭档智能体截图诊断接入 Vision 多模态，注册 front-end-agent.assistant::vision |
 | fix | .claude/skills/cds | cdscli preview-url 本地推算优先 git remote 仓库名，修复 Cloud Agent workspace 目录拼错预览域 |
